Restructure sessions package in core
COSDK-453
