+++ core/src/main/java/com/adyen/checkout/core/internal/ui/model/SessionParamsFactory.kt
@@ -11,8 +11,8 @@ package com.adyen.checkout.core.internal.ui.model
 import androidx.annotation.RestrictTo
 import com.adyen.checkout.core.internal.util.LocaleUtil
 import com.adyen.checkout.core.sessions.CheckoutSession
-import com.adyen.checkout.core.sessions.data.model.SessionDetails
-import com.adyen.checkout.core.sessions.data.model.mapToDetails
+import com.adyen.checkout.core.sessions.internal.data.model.SessionDetails
+import com.adyen.checkout.core.sessions.internal.data.model.mapToDetails
 import java.util.Locale
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
