+++ core/src/main/java/com/adyen/checkout/core/sessions/internal/data/api/SessionRepository.kt
@@ -3,20 +3,20 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by ozgur on 30/4/2025.
+ * Created by ozgur on 12/5/2025.
  */
 
-package com.adyen.checkout.core.sessions.data.api
+package com.adyen.checkout.core.sessions.internal.data.api
 
 import androidx.annotation.RestrictTo
 import com.adyen.checkout.core.data.PaymentComponentData
 import com.adyen.checkout.core.internal.util.runSuspendCatching
 import com.adyen.checkout.core.paymentmethod.PaymentMethodDetails
 import com.adyen.checkout.core.sessions.SessionModel
-import com.adyen.checkout.core.sessions.data.SessionPaymentsRequest
-import com.adyen.checkout.core.sessions.data.SessionPaymentsResponse
-import com.adyen.checkout.core.sessions.data.SessionSetupRequest
-import com.adyen.checkout.core.sessions.data.SessionSetupResponse
+import com.adyen.checkout.core.sessions.internal.data.model.SessionPaymentsRequest
+import com.adyen.checkout.core.sessions.internal.data.model.SessionPaymentsResponse
+import com.adyen.checkout.core.sessions.internal.data.model.SessionSetupRequest
+import com.adyen.checkout.core.sessions.internal.data.model.SessionSetupResponse
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class SessionRepository(
