+++ core/src/main/java/com/adyen/checkout/core/sessions/internal/data/model/SessionPaymentsRequest.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 12/5/2025.
  */
 
-package com.adyen.checkout.core.sessions.data
+package com.adyen.checkout.core.sessions.internal.data.model
 
 import androidx.annotation.RestrictTo
 import com.adyen.checkout.core.data.PaymentComponentData
